requirements.txt
streamlit
pandas
openpyxl
scikit-learn
matplotlib
seaborn
plotly==5.19.0
python-dotenv
openai>=1.0.0
datetime